Moved the generic data display code from ViewResultsFrame to the view_results_mixin module
